Time: 0 ms (100.00%), Space: 7.4 MB (52.08%) - LeetHub

+++ 0007-reverse-integer/0007-reverse-integer.cpp
@@ -1,7 +1,8 @@
 class Solution {
 public:
     int reverse(int x) {
-        long long reversedNum;
+    long long reversedNum;
+/*Consider the case where the input x is a large number, such as 1534236469. When you reverse this number, you get 9646324351, which exceeds the maximum value representable by a 32-bit signed integer (int). Storing this value in an int would result in integer overflow, leading to undefined behavior.*/
         reversedNum=0;
     if(x>0)
     {while(x>0)
@@ -22,7 +23,7 @@ public:
         if (reversedNum > INT_MAX || reversedNum < INT_MIN) {
             return 0; // Return 0 for overflow
         } else {
-            return static_cast<int>(reversedNum);
+            return reversedNum;
         }
     }
 };
